Try moving parts for distributed tests to tests/conftest.py
